refactor confirm position to be its own step
