+++ app/src/organisms/DropTipWizardFlows/constants.ts
@@ -6,6 +6,7 @@ export const BLOWOUT_SUCCESS = 'BLOWOUT_SUCCESS' as const
 export const CHOOSE_BLOWOUT_LOCATION = 'CHOOSE_BLOWOUT_LOCATION' as const
 export const CHOOSE_DROP_TIP_LOCATION = 'CHOOSE_DROP_TIP_LOCATION' as const
 export const POSITION_AND_DROP_TIP = 'POSITION_AND_DROP_TIP' as const
+export const CONFIRM_POSITION = 'CONFIRM_POSITION' as const
 export const DROP_TIP_SUCCESS = 'DROP_TIP_SUCCESS' as const
 export const INVALID = 'INVALID' as const
 
@@ -13,11 +14,13 @@ export const BEFORE_BEGINNING_STEPS = [BEFORE_BEGINNING] as const
 export const BLOWOUT_STEPS = [
   CHOOSE_BLOWOUT_LOCATION,
   POSITION_AND_BLOWOUT,
+  CONFIRM_POSITION,
   BLOWOUT_SUCCESS,
 ] as const
 export const DROP_TIP_STEPS = [
   CHOOSE_DROP_TIP_LOCATION,
   POSITION_AND_DROP_TIP,
+  CONFIRM_POSITION,
   DROP_TIP_SUCCESS,
 ] as const
