+++ app/src/organisms/DropTipWizardFlows/steps/index.ts
@@ -2,3 +2,4 @@ export * from './BeforeBeginning'
 export * from './ChooseLocation'
 export * from './JogToPosition'
 export * from './Success'
+export * from './ConfirmPosition'
